edit test

+++ SamulNori_Library/src/main/java/com/sample/cep/SmokeSensorData.java
@@ -1,14 +1,14 @@
 package com.sample.cep;
 
 public class SmokeSensorData extends SensorData {
-	public SmokeSensorData(double value) {
+	public SmokeSensorData(long value) {
 		setSmokeLevel(value);
 		setWriteValue(1<<1);
 	}
 	public double getSmokeLevel() {
 		return getDoubleValue();
 	}
-	public void setSmokeLevel(double value) {
+	public void setSmokeLevel(long value) {
 		setDoubleValue(value);
 	}
 	public long getSmokeLimit()
